added GO on all sql files, fixed AP SP drop procedure catch

+++ Integration/dbo/Stored Procedures/uspAPImportBillTransactions.sql
@@ -1,4 +1,6 @@
-﻿IF EXISTS(select top 1 1 from INFORMATION_SCHEMA.VIEWS where TABLE_NAME = 'uspAPImportBillTransactions')
+﻿GO
+
+IF EXISTS(select top 1 1 from INFORMATION_SCHEMA.VIEWS where TABLE_NAME = 'uspAPImportBillTransactions')
 	DROP VIEW uspAPImportBillTransactions
 
 IF  (SELECT TOP 1 ysnUsed FROM ##tblOriginMod WHERE strPrefix = 'AP') = 1
