+++ Integration/dbo/Views/apcbkmst.sql
@@ -1,4 +1,7 @@
-﻿IF EXISTS(select top 1 1 from INFORMATION_SCHEMA.VIEWS where TABLE_NAME = 'apcbkmst')
+﻿GO
+
+
+IF EXISTS(select top 1 1 from INFORMATION_SCHEMA.VIEWS where TABLE_NAME = 'apcbkmst')
 	DROP VIEW apcbkmst
 
 IF  (SELECT TOP 1 ysnUsed FROM ##tblOriginMod WHERE strPrefix = 'AP') = 1
@@ -10,3 +13,6 @@ BEGIN
 		SELECT * FROM apcbkmst_origin
 	')
 END
+
+GO
+
